Added a GPL2 LICENSE
Update added a GPL2 LICENSE
Added this license, so there would be some license, for making the repo public.

I intend to discuss the final license with stakeholders, so a suitable, minimal hassle solution for ensuring a healthy future for this project.

If for any reason someone finds the license too restrictive for what they want to do, I'm open granting individual licenses for those who need it.

My hope is that this project would be useful for as many people as possible, and to stay useful, any improvements made would be contributed back to the project. 

Any images, CSS styling or text content can be changed without releasing those assets.