Removed ineffective "raw" maps
The idea was that the pointers would point to the item in map,
but this is not the case. It turns out there is no way to implement
this in Go, as such the raw map was removed and replaced with
another "byID" map.

To edit an item, a pointer is the easiest and cleanest way. A host
structure is not needed for the real object. Pointers are the way.