Removed dead code & Added server side logging for failed auth events
Removed unnecessary references to plaintext fingerprint.
Fingerprint is used only in resuming an existing session, where as
HashPrint is the securely hashed version of it and used when
creating a new session.

As the hash algorythm uses a random salt on each hash, we need to
use the provided function to compare tha bare fingerprint to the
hash, not create a new hash.